Separate code_statistics into calculator and formatter
This commit adds a new formatter (JSON) and it refactors CodeStatistics to coordinate between calculator and formatter.

The goal is to be able to call the rake stats with a parameter that will output the result in JSON format.
