+++ lib/rails_stats/stats_formatter.rb
@@ -1,0 +1,11 @@
+module RailsStats
+  class StatsFormatter
+    def initialize(calculator, opts = {})
+      @calculator = calculator
+      @statistics = calculator.statistics
+      @total      = calculator.total
+    end
+
+    attr_reader :calculator
+  end
+end
